+++ backend/services/GatherService/Controllers/AccountsController.cs
@@ -1,0 +1,42 @@
+﻿using Asp.Versioning;
+using Gather.Dtos;
+using Gather.Models;
+using Gather.Services.AccountService;
+using Microsoft.AspNetCore.Mvc;
+using SharedCore.Filtering;
+
+namespace Gather.Controllers
+{
+    [ApiController]
+    [Produces("application/json")]
+    [Route("api/v{version:apiVersion}/[controller]")]
+    [ApiVersion("1.0")]
+    public class AccountsController:ControllerBase
+    {
+        private readonly IAccountService _accountService;
+
+        public AccountsController(IAccountService accountService)
+        {
+            _accountService = accountService;
+        }
+
+        [HttpGet("{accountTlgId}")]
+        [MapToApiVersion(1.0)]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        public async Task<ActionResult<ServiceResponse<AccountDto>>> GetSingle([FromRoute] long accountTlgId)
+        {
+            var response = await _accountService.GetAccountAsync(accountTlgId);
+            if (response.Data == null)
+            {
+                return NotFound(response);
+            }
+            else if (response.Success == false)
+            {
+                return StatusCode(StatusCodes.Status500InternalServerError, response);
+            }
+            return Ok(response);
+        }
+    }
+}
