+++ backend/services/GatherService/Dtos/LoginDto.cs
@@ -1,9 +1,10 @@
-﻿namespace Gather.Dtos
+﻿namespace Gather.Dtos;
+
+public class LoginDto
 {
-    public class LoginDto
-    {
-        public string UserName { get; set; } = string.Empty;
-        public string Password { get; set; } = string.Empty;
-        public string PhoneNumber { get; set; } = string.Empty;
-    }
+    public string UserName { get; set; } = string.Empty;
+
+    public string Password { get; set; } = string.Empty;
+
+    public string PhoneNumber { get; set; } = string.Empty;
 }
