+++ backend/services/GatherService/Dtos/User/UserLoginDto.cs
@@ -1,0 +1,8 @@
+﻿namespace Gather.Dtos;
+
+public class UserLoginDto
+{
+    public string Username { get; set; } = string.Empty;
+    public string Password { get; set; } = string.Empty;
+    public string PhoneNumber { get; set; }=string.Empty;
+}
