+++ backend/services/GatherService/Dtos/User/UserRegisterDto.cs
@@ -1,0 +1,7 @@
+﻿namespace Gather.Dtos;
+
+public class UserRegisterDto
+{
+    public string Username { get; set; } = string.Empty;
+    public string Password { get; set; } = string.Empty;
+}
