+++ backend/services/GatherService/Models/ServiceResponse.cs
@@ -1,0 +1,10 @@
+﻿namespace Gather.Models;
+
+public class ServiceResponse<T>
+{
+    public T? Data { get; set; }
+
+    public bool Success { get; set; } = true;
+
+    public string Message { get; set; } = string.Empty;
+}
