+++ backend/services/GatherService/Services/AccountService/IAccountService.cs
@@ -1,0 +1,9 @@
+﻿using Gather.Dtos;
+using Gather.Models;
+
+namespace Gather.Services.AccountService;
+
+public interface IAccountService
+{
+    Task<ServiceResponse<AccountDto>> GetAccountAsync(long accountTlgId);
+}
