+++ backend/services/GatherService/Services/GatherService/IGatherService.cs
@@ -1,12 +1,12 @@
 ﻿using Gather.Models;
-using SharedCore.Models;
 
-namespace Gather.Services
+namespace Gather.Services;
+
+public interface IGatherService
 {
-    public interface IGatherService
-    {
-        Task<ServiceResponse<bool>> StartGatherAllAsync(string username);
-        Task<ServiceResponse<bool>> GetGatherStatusAsync(string username);
-        Task<ServiceResponse<bool>> StopGatherStatusAsync(string username);
-    }
+    Task<ServiceResponse<bool>> StartGatherAllAsync(string username);
+
+    Task<ServiceResponse<bool>> GetGatherStatusAsync(string username);
+    
+    Task<ServiceResponse<bool>> StopGatherStatusAsync(string username);
 }
