+++ backend/services/GatherService/Services/UserService/IUserService.cs
@@ -1,14 +1,17 @@
-﻿using SharedCore.Dtos.User;
-using SharedCore.Models;
+﻿using Gather.Dtos;
+using Gather.Models;
 
-namespace Gather.Services.UserService
+namespace Gather.Services.UserService;
+
+public interface IUserService
 {
-    public interface IUserService
-    {
-        Task<ServiceResponse<IEnumerable<GetUserDto>>> GetAllUsersAsync();
-        Task<ServiceResponse<GetUserDto>> GetUserIdAsync(int id);
-        Task<ServiceResponse<int>> AddUserAsync(AddUserDto newUser);
-        Task<ServiceResponse<GetUserDto>> DeleteUserAsync(int id);
-        Task<ServiceResponse<GetUserDto>> EditUserAsync(User user);
-    }
+    Task<ServiceResponse<IEnumerable<GetUserDto>>> GetAllUsersAsync();
+
+    Task<ServiceResponse<GetUserDto>> GetUserIdAsync(int id);
+
+    Task<ServiceResponse<int>> AddUserAsync(AddUserDto newUser);
+
+    Task<ServiceResponse<GetUserDto>> DeleteUserAsync(int id);
+
+    Task<ServiceResponse<GetUserDto>> EditUserAsync(User user);
 }
