--- backend/services/SharedCore/Dtos/PostDto.cs
@@ -1,14 +1,0 @@
-﻿using SharedCore.Models;
-
-namespace SharedCore.Dtos
-{
-    public class PostDto
-    {
-        public long TlgId { get; set; }
-        public long PeerId { get; set; }
-        public Account? Author { get; set; }
-        public string? Message { get; set; }
-        public DateTime Date { get; set; }
-        public long CommentsCount { get; set; }
-    }
-}
