--- backend/services/SharedCore/Dtos/User/GetUserDto.cs
@@ -1,10 +1,0 @@
-﻿namespace SharedCore.Dtos.User
-{
-    public class GetUserDto
-    {
-        public int UserId { get; set; }
-        public string? Username { get; set; }
-        public string? Password { get; set; }
-        public string? PhoneNumber { set; get; }
-    }
-}
