--- backend/services/SharedCore/Dtos/User/UserLoginDto.cs
@@ -1,9 +1,0 @@
-﻿namespace SharedCore.Dtos.User
-{
-    public class UserLoginDto
-    {
-        public string Username { get; set; } = string.Empty;
-        public string Password { get; set; } = string.Empty;
-        public string PhoneNumber { get; set; }=string.Empty;
-    }
-}
