--- backend/services/SharedCore/Dtos/User/UserRegisterDto.cs
@@ -1,8 +1,0 @@
-﻿namespace SharedCore.Dtos.User
-{
-    public class UserRegisterDto
-    {
-        public string Username { get; set; } = string.Empty;
-        public string Password { get; set; } = string.Empty;
-    }
-}
