--- backend/services/SharedCore/Models/ServiceResponse.cs
@@ -1,11 +1,0 @@
-﻿using System.Net;
-
-namespace SharedCore.Models
-{
-    public class ServiceResponse<T>
-    {
-        public T? Data { get; set; }
-        public bool Success { get; set; } = true;
-        public string Message { get; set; } = String.Empty;
-    }
-}
